Extract ClusterMergeTask from ClusterService
ClusterMergeTask is triggered on every member in the cluster when master detects
another cluster to join.
